Changed "Log In" on the navbar to "Login"
"Login" is a noun and thus it is appropriate to use on the navbar because it is describing a location, the login page. "Log In" is a verb, which is why it is used on the actual "Log in" button to describe the action of logging in. See also: http://english.stackexchange.com/questions/23178/login-or-log-in